Add files via upload
Added some extra colors for the new categories. The new math looks a the most important areas first like food and adds a max of 100 to the budget. If there is more than 100 left it adds 100 to food and then moves to the next category. If there is not $100 left after rent etc it adds what is left to food and sets the others to 0. So there may be times it sets half of them to a number and the rest to 0 or everything but savings gets some money. Figured this would be best since you dont want to have 4 or 5 for internet etc